+++ geotrek/core/models.py
@@ -157,6 +157,12 @@ class Path(
     objects = PathManager()
     include_invisible = PathInvisibleManager()
 
+    is_being_split = models.BooleanField(
+        default=False,
+        editable=False,
+        help_text="Internal field preventing geometry updates on related topologies during path splitting",
+    )
+
     is_reversed = False
     can_duplicate = False
 
@@ -167,19 +173,6 @@ class Path(
     @classproperty
     def length_2d_verbose_name(cls):
         return _("2D Length")
-
-    @classmethod
-    def no_draft_latest_updated(cls):
-        try:
-            latest = (
-                cls.objects.filter(draft=False)
-                .only("date_update")
-                .latest("date_update")
-                .get_date_update()
-            )
-        except cls.DoesNotExist:
-            latest = None
-        return latest
 
     @property
     def length_2d_display(self):
@@ -411,42 +404,37 @@ class Path(
         return ", ".join([str(n) for n in self.networks.all()])
 
     def topologies_by_path(self, default_dict):
-        if "geotrek.core" in settings.INSTALLED_APPS:
-            for trail in self.trails:
-                default_dict[_("Trails")].append(
-                    {"name": trail.name, "url": trail.get_detail_url()}
-                )
-        if "geotrek.trekking" in settings.INSTALLED_APPS:
-            for trek in self.treks:
-                default_dict[_("Treks")].append(
-                    {"name": trek.name, "url": trek.get_detail_url()}
-                )
-            for service in self.services:
-                default_dict[_("Services")].append(
-                    {"name": service.type.name, "url": service.get_detail_url()}
-                )
-            for poi in self.pois:
-                default_dict[_("Pois")].append(
-                    {"name": poi.name, "url": poi.get_detail_url()}
-                )
-        if "geotrek.signage" in settings.INSTALLED_APPS:
-            for signage in self.signages:
-                default_dict[_("Signages")].append(
-                    {"name": signage.name, "url": signage.get_detail_url()}
-                )
-        if "geotrek.infrastructure" in settings.INSTALLED_APPS:
-            for infrastructure in self.infrastructures:
-                default_dict[_("Infrastructures")].append(
-                    {
-                        "name": infrastructure.name,
-                        "url": infrastructure.get_detail_url(),
-                    }
-                )
-        if "geotrek.maintenance" in settings.INSTALLED_APPS:
-            for intervention in self.interventions:
-                default_dict[_("Interventions")].append(
-                    {"name": intervention.name, "url": intervention.get_detail_url()}
-                )
+        for trail in self.trails:
+            default_dict[_("Trails")].append(
+                {"name": trail.name, "url": trail.get_detail_url()}
+            )
+        for trek in self.treks:
+            default_dict[_("Treks")].append(
+                {"name": trek.name, "url": trek.get_detail_url()}
+            )
+        for service in self.services:
+            default_dict[_("Services")].append(
+                {"name": service.type.name, "url": service.get_detail_url()}
+            )
+        for poi in self.pois:
+            default_dict[_("Pois")].append(
+                {"name": poi.name, "url": poi.get_detail_url()}
+            )
+        for signage in self.signages:
+            default_dict[_("Signages")].append(
+                {"name": signage.name, "url": signage.get_detail_url()}
+            )
+        for infrastructure in self.infrastructures:
+            default_dict[_("Infrastructures")].append(
+                {
+                    "name": infrastructure.name,
+                    "url": infrastructure.get_detail_url(),
+                }
+            )
+        for intervention in self.interventions:
+            default_dict[_("Interventions")].append(
+                {"name": intervention.name, "url": intervention.get_detail_url()}
+            )
 
     def merge_path(self, path_to_merge):
         """
@@ -515,6 +503,12 @@ class Topology(
         verbose_name=_("Length 2D"),
     )
     uuid = models.UUIDField(default=uuid.uuid4, editable=False, unique=True)
+
+    # The geometry is coupled to the path network when the topology is valid and linked to path aggregations
+    # See https://github.com/GeotrekCE/Geotrek-admin/issues/4982
+    coupled = models.BooleanField(
+        default=False, editable=False, verbose_name=_("Network-coupled")
+    )
 
     """ Fake srid attribute, that prevents transform() calls when using Django map widgets. """
     srid = settings.API_SRID
@@ -648,17 +642,13 @@ class Topology(
 
     def mutate(self, other):
         """
-        Take alls attributes of the other topology specified and
-        save them into this one. Optionnally deletes the other.
+        Take all attributes of the other topology specified and save them into this one.
         """
         self.offset = other.offset
         self.save(update_fields=["offset"])
         PathAggregation.objects.filter(topo_object=self).delete()
-        # The previous operation has put deleted = True (in triggers)
-        # and NULL in geom (see update_geometry_of_topology:: IF t_count = 0)
-        self.deleted = False
         self.geom = other.geom
-        self.save(update_fields=["deleted", "geom"])
+        self.save(update_fields=["geom"])
 
         # Now copy all agregations from other to self
         aggrs = other.aggregations.all()