+++ geotrek/signage/forms.py
@@ -156,8 +156,8 @@ class SignageForm(BaseSignageForm):
         Div(
             "structure",
             "name",
-            "description",
             "type",
+            "description",
             "conditions",
             "implantation_year",
             "published",