+++ docs/changelog.rst
@@ -5,8 +5,22 @@ CHANGELOG
 2.124.2+dev     (XXXX-XX-XX)
 ----------------------------
 
+**Warnings**
+
+* Topologies no longer need to depend on the path network: they can now exist independently or be based on paths, regardless of the dynamic segmentation setting.
+* All geotrek modules are now installed by default. Please remove `geotrek.outdoor`, `geotrek.sensitivity`, `geotrek.diving` from you custom.py file. User setting to disable module if you don't want to use them.
+  - `COURSE_MODEL_ENABLED = FALSE`
+  - `SITE_MODEL_ENABLED = FALSE`  # for outdoor module
+  - `SENSITIVE_AREA_MODEL_ENABLED = FALSE`  # for sensitivity module
+  - `DIVE_MODEL_ENABLED = FALSE`  # for diving module
+
 **Improvements**
 
+* Add a "Network-coupled" boolean field and its corresponding filter to the ``Topology`` model
+* Don't update the geometry of a topology when it's decoupled from the path network
+* Strengthen the "invalid topology" filter criteria
+* Remove the "invalid geometry" filter
+* From now, if a module is disabled by its setting, it is completely hidden from menu, layer tree, admin and API responses.
 * Add ``--ids`` option to the ``reorder_topologies`` command to process specific topologies by ID
 * Remove blade double link (keep signage ForeignKey and remove topology ForeignKey) (#3009)
 
@@ -17,6 +31,7 @@ CHANGELOG
 **Documentation**
 
 * Add information about database users and map style configuration, and remove the mention of IGN Rando (refs #5318)
+
 
 2.124.1         (2026-02-17)
 ----------------------------